Move errors and error utils to a shared component.
(cherry picked from commit da4f91b94280236b209346a127d645a8a8f97ccb)
